requirements.txt
streamlit
folium
streamlit-folium
pymongo
python-dotenv